Use busybox_defconfig and follow best practises and upstream rec.
Use myLinux's busybox_defconfig as base for the configuration.  However,
not all features BusyBox use work properly with musl libc.  The latter
is a strict POSIX C library and as such it does not have BSD extensions
like REG_STARTEND, for instance.

To build we need som extra hand waving, which the Buildroot project have
so nicely documented in packages/pkg-kconfig.mk.  This because BusyBox
has an older kconfig implementation and doesn't even have oldnoconfig.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>